Enable multidex for android pre-5.0
