+++ android/app/build.gradle
@@ -45,6 +45,7 @@ android {
         applicationId "co.banano.natriumwallet"
         minSdkVersion 18
         targetSdkVersion 28
+        multiDexEnabled true
         versionCode flutterVersionCode.toInteger()
         versionName flutterVersionName
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -94,4 +95,5 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.3.11"
     implementation "com.bottlerocketstudios:vault:1.4.2"
+    implementation 'androidx.multidex:multidex:2.0.0'
 }
