+++ android/app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
          additional functionality it is fine to subclass or reimplement
          FlutterApplication and put your custom class here. -->
     <application
-        android:name="io.flutter.app.FlutterApplication"
+        android:name=".MultidexApplication"
         android:label="Natrium"
         android:icon="@mipmap/ic_launcher">
         <meta-data
